refactor : 게임 상태 패키지 이름 변경 (State)
