+++ src/main/java/chess/controller/state/CommandType.java
@@ -1,4 +1,4 @@
-package chess.controller.status;
+package chess.controller.state;
 
 public enum CommandType {
     START, MOVE, END
