+++ src/main/java/chess/controller/state/State.java
@@ -1,0 +1,9 @@
+package chess.controller.state;
+
+import chess.controller.Command;
+
+public interface State {
+    State checkCommand(final Command command);
+
+    boolean isRun();
+}
