--- src/main/java/chess/controller/status/Status.java
@@ -1,9 +1,0 @@
-package chess.controller.status;
-
-import chess.controller.Command;
-
-public interface Status {
-    Status checkCommand(final Command command);
-
-    boolean isRun();
-}
