+++ src/test/java/chess/controller/state/StartTest.java
@@ -1,4 +1,4 @@
-package chess.controller.status;
+package chess.controller.state;
 
 import chess.controller.Command;
 import chess.domain.chess.ChessGame;
@@ -33,10 +33,10 @@ class StartTest {
         final Command command = new Command(CommandType.END, List.of("end"));
 
         // when
-        Status status = start.checkCommand(command);
+        State state = start.checkCommand(command);
 
         // then
-        assertThat(status)
+        assertThat(state)
                 .isInstanceOf(End.class);
     }
